+++ src/components/form-sample/configured-form-sample.jsx
@@ -3,7 +3,7 @@ import _throttle from 'lodash/throttle';
 
 import FormSample from './form-sample.component';
 import validate from './form-sample-validations';
-import { asyncValidateDatasetInvokePatch } from '../formsLib/asyncValidateDatasetInvokePatch';
+import { asyncValidateDatasetInvokePatch } from '../../pages/dataset-registration-page/formsLib/asyncValidateDatasetInvokePatch';
 
 const config = {
   form: 'sample',
