+++ src/components/form-sample/form-sample.component.tsx
@@ -5,18 +5,18 @@ import { Field, FieldArray } from 'redux-form';
 import {
   withTranslations,
   Props as TranslationsProps
-} from '../../../providers/translations';
+} from '../../providers/translations';
 
-import Translation from '../../../components/translation';
-import Helptext from '../../../components/helptext/helptext.component';
-import InputField from '../../../components/fields/field-input/field-input.component';
-import MultilingualField from '../../../components/multilingual-field/multilingual-field.component';
-import InputTagsField from '../../../components/fields/field-input-tags/field-input-tags.component';
-import TextAreaField from '../../../components/fields/field-textarea/field-textarea.component';
-import { licenseType, textType } from '../../../schemaTypes';
-import { datasetFormPatchThunk } from '../formsLib/asyncValidateDatasetInvokePatch';
-import LinkReadonlyField from '../../../components/fields/field-link-readonly/field-link-readonly.component';
-import InputFieldReadonly from '../../../components/fields/field-input-readonly/field-input-readonly.component';
+import Translation from '../translation';
+import Helptext from '../helptext/helptext.component';
+import InputField from '../fields/field-input/field-input.component';
+import MultilingualField from '../multilingual-field/multilingual-field.component';
+import InputTagsField from '../fields/field-input-tags/field-input-tags.component';
+import TextAreaField from '../fields/field-textarea/field-textarea.component';
+import { licenseType, textType } from '../../schemaTypes';
+import { datasetFormPatchThunk } from '../../pages/dataset-registration-page/formsLib/asyncValidateDatasetInvokePatch';
+import LinkReadonlyField from '../fields/field-link-readonly/field-link-readonly.component';
+import InputFieldReadonly from '../fields/field-input-readonly/field-input-readonly.component';
 
 interface ExternalProps {
   dispatch: (arg: any) => void;
