+++ src/components/form-sample/form-sample.test.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { shallow } from 'enzyme';
 import FormSample from './form-sample.component';
 import { sampleTypes } from './connected-form-sample.component';
-import openlicenses from '../../../../test/fixtures/openlicenses';
-import samples from '../../../../test/fixtures/samples';
+import openlicenses from '../../../test/fixtures/openlicenses';
+import samples from '../../../test/fixtures/samples';
 
 let defaultProps;
 let wrapper;
